requirements.txt
flask==3.0.0
twilio==8.10.0
google-generativeai==0.3.2
python-dotenv==1.0.0
Flask-Caching==2.1.0
gunicorn==21.2.0
flask-talisman==0.8.1
werkzeug==2.3.7
mysql-connector-python==8.3.0